Update Model
- Convert answers.qualities to competences
- Update candidate competences

+++ Assets/Scripts/AIengine/M_Competence.cs
@@ -9,10 +9,10 @@ namespace HRAP
     {
         private string name;
         private bool isTech;
-        private int points;
+        private double points;
         private bool isImportant;
 
-        public M_Competence(string name, bool isTech, int points, bool isImportant)
+        public M_Competence(string name, bool isTech, double points, bool isImportant)
         {
             this.name = name;
             this.isTech = isTech;
@@ -22,7 +22,7 @@ namespace HRAP
 
         public string Name { get { return name; } }
         public bool IsTech { get { return isTech; } }
-        public int Points { get { return points; } set { points = value; } }
+        public double Points { get { return points; } set { points = value; } }
         public bool IsImportant { get { return isImportant; } set { isImportant = value; } }
 
     }
